requirements.txt
streamlit==1.32.0
transformers==4.38.2
torch==2.2.1
gtts==2.5.1
pydub==0.25.1
nltk==3.8.1
vaderSentiment==3.3.2
textblob==0.18.0
numpy==1.26.4
scipy==1.12.0
pyttsx3==2.90
requests==2.31.0
plotly==5.19.0
python-dotenv==1.0.1
